components refactoring
