+++ src/components/AsideBar/AsideBar.module.scss
@@ -61,42 +61,6 @@
     font-size: 2.5rem;
   }
 
-  &__task {
-    display: flex;
-    flex-direction: column;
-    align-items: start;
-    justify-content: space-between;
-
-    margin-bottom: 1.5rem;
-  }
-
-  &__content {
-    display: flex;
-    flex-direction: row;
-    align-items: start;
-    text-align: left;
-
-    gap: 1rem;
-
-    font-size: 1.5rem;
-
-    margin: 1rem;
-    padding: 1rem 1rem 1rem 0.7rem;
-
-    &Color{
-      position: inherit;
-
-      min-width: 1.5rem;
-      min-height: 1.5rem;
-
-      border-radius: 50%;
-
-      box-shadow: 0 0 10px var(--defaultShadowColor);
-
-      margin-top: 0.4rem;
-    }
-  }
-
   &__addButtonArea {
     font-size: 3rem;
 
@@ -107,15 +71,5 @@
 
       transition: 1s;
     }
-  }
-
-  &__deleteEditButtonArea {
-    display: flex;
-    flex-direction: row;
-
-    gap: 1rem;
-
-    margin-top: -1.3rem;
-    margin-left: 3.5rem;
   }
 }
