+++ src/components/Calendar/Calendar.module.scss
@@ -111,29 +111,4 @@
 
     padding: 1rem;
   }
-
-  &__tasks{
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-
-    overflow: auto;
-    &::-webkit-scrollbar{
-      border: none;
-
-      &-thumb{
-        border: none;
-      }
-    }
-
-    gap: 1rem;
-  }
-
-  &__task {
-    color: rgba(0, 0, 0, 0.65);
-
-    padding: 0.35rem;
-
-    border-radius: 1rem;
-  }
 }
