+++ src/components/Calendar/SupportComponents/ThemeSelector/ThemeSelector.module.scss
@@ -1,0 +1,12 @@
+.DarkSelector {
+  cursor: pointer;
+
+  margin-top: 3.5rem;
+}
+
+.LightSelector {
+  cursor: pointer;
+  color: white;
+
+  margin-top: 3.5rem;
+}
